requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
audiodiffusion==1.5.6
audioread==3.0.1
backcall==0.2.0
bokeh==3.1.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
comm==0.1.4
contourpy==1.1.1
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
diffusers==0.21.4
dill==0.3.7
einops==0.7.0
executing==2.0.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.6.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.12.3
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter-client==8.4.0
jupyter-core==5.4.0
jupyterlab-widgets==3.0.9
-e git+https://github.com/CompVis/stable-diffusion.git@21f890f9da3cfbeaba8e2ac3c425ee9e998d5229#egg=latent_diffusion
lazy-loader==0.3
librosa==0.10.1
lightning-utilities==0.9.0
llvmlite==0.41.1
Markdown==3.5
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pooch==1.7.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.16.1
python-dateutil==2.8.2
pytorch-lightning==1.7.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
sympy==1.12
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==1.13.1+cu117
torch-tb-profiler==0.4.3
torchaudio==0.13.1+cu117
torchmetrics==0.11.4
torchvision==0.14.1+cu117
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
werkzeug==3.0.0
widgetsnbextension==4.0.9
xxhash==3.4.1
xyzservices==2023.10.1
yarl==1.9.2
zipp==3.17.0
